Move the arm to it's own subsystem
