+++ src/main/java/frc/robot/Constants.java
@@ -23,6 +23,28 @@ public final class Constants {
     public static final double JOYSTICK_DEADBAND = 0.05;
 
     /**
+     * Constants for the arm
+     */
+    public final static class Arm {
+        /******** Motor CAN ID's ********/
+        public static final int ARM_MOTOR_ID = 8;
+
+        /******** PID Constants ********/
+        public static final PIDParameters ARM_PID = new PIDParameters(0.2, 2.54972071e-05, 10.0); // Gains taken from 2020 robot
+
+        /******** Physical Constants ********/
+        public static final int LIMIT_SWITCH_PORT = 0; // Digital input port on the RoboRIO that the arm's limit switch is connected too
+        public static final double LIMIT_SWITCH_ANGLE = -10; // The angle that the arm is at when the limit switch is triggered
+
+        public static final double ARM_GEAR_RATIO = 45.0 / 1;
+
+        /******** Arm Location Presets ********/
+        public static final double START_CONFIG_ANGLE = 30; // Angle where robot fits in starting config
+        public static final double STOWED_ANGLE = -5; // Angle where robot fits under the Wheel of Fortune (WOF)
+        public static final double INTAKE_ANGLE = -10; // Angle required for intaking power cells
+    }
+
+    /**
      * Constants for the drivetrain
      */
     public final static class Drive {
@@ -75,24 +97,11 @@ public final class Constants {
     public final static class Shooter {
         /******** Motor CAN ID's ********/
         public static final int FLYWHEEL_MOTOR_ID = 7;
-        public static final int ARM_MOTOR_ID = 8;
 
         /******** PID Constants ********/
         public static final PIDParameters FLYWHEEL_PID = new PIDParameters(0.1, 0, 0);
-        public static final PIDParameters ARM_PID = new PIDParameters(0.2, 2.54972071e-05, 10.0); // Gains taken from 2020 robot
 
         /******** Tuning Constants ********/
         public static final int FLYWHEEL_SPEED_TOLERANCE = 50; // Only shoot powercells if flywheels are within this range of their target speed
-
-        /******** Physical Constants ********/
-        public static final int LIMIT_SWITCH_PORT = 0; // Digital input port on the RoboRIO that the arm's limit switch is connected too
-        public static final double LIMIT_SWITCH_ANGLE = -10; // The angle that the arm is at when the limit switch is triggered
-
-        public static final double ARM_GEAR_RATIO = 45.0 / 1;
-
-        /******** Arm Location Presets ********/
-        public static final double START_CONFIG_ANGLE = 30; // Angle where robot fits in starting config
-        public static final double STOWED_ANGLE = -5; // Angle where robot fits under the Wheel of Fortune (WOF)
-        public static final double INTAKE_ANGLE = -10; // Angle required for intaking power cells
     }
 }
