+++ src/main/java/frc/robot/subsystems/ShooterSubsystem.java
@@ -4,7 +4,6 @@ import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonFX;
 
-import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.UnitConversions;
@@ -13,27 +12,17 @@ import friarLib2.utility.PIDParameters;
 /**
  * Represents the shooter subsystem
  * 
- * <p>The physical subsystem includes the flywheels and arm
+ * <p>The physical subsystem includes the flywheels
  */
 public class ShooterSubsystem extends SubsystemBase {
 
     private WPI_TalonFX flywheelMotor;
-    private WPI_TalonFX armMotor;
-    private DigitalInput limitSwitch;
 
     public ShooterSubsystem () {
         flywheelMotor = new WPI_TalonFX(Constants.Shooter.FLYWHEEL_MOTOR_ID);
         flywheelMotor.configFactoryDefault();
         flywheelMotor.setNeutralMode(NeutralMode.Coast);
         PIDParameters.configureMotorPID(flywheelMotor, Constants.Shooter.FLYWHEEL_PID);
-
-        armMotor = new WPI_TalonFX(Constants.Shooter.ARM_MOTOR_ID);
-        armMotor.configFactoryDefault();
-        armMotor.setNeutralMode(NeutralMode.Brake);
-        PIDParameters.configureMotorPID(armMotor, Constants.Shooter.ARM_PID);
-
-        limitSwitch = new DigitalInput(Constants.Shooter.LIMIT_SWITCH_PORT);
-        zeroArmPosition();
     }
 
     /**
@@ -51,33 +40,6 @@ public class ShooterSubsystem extends SubsystemBase {
 
     public void stopFlywheel () {
         flywheelMotor.stopMotor();
-    }
-
-    /**
-     * Set the position of the arm
-     * 
-     * <p>When the aluminum structure of the arm is horizontal to the
-     * ground, the arm is at zero degrees.
-     * 
-     * <p>The degrees increase as the arm moves closer to vertical.
-     * 
-     * @param degrees
-     */
-    public void setArmPosition (double degrees) {
-        armMotor.set(ControlMode.Position, UnitConversions.Shooter.armDegreesToEncoderTicks(degrees));
-    }
-
-    /**
-     * Zero the arm's encoder
-     * 
-     * <p>Assumes that the arm is at the limit switch
-     */
-    public void zeroArmPosition () {
-        armMotor.setSelectedSensorPosition(UnitConversions.Shooter.armDegreesToEncoderTicks(Constants.Shooter.LIMIT_SWITCH_ANGLE));
-    }
-
-    public double getArmPosition () {
-        return UnitConversions.Shooter.armEncoderTicksToDegrees(armMotor.getSelectedSensorPosition());
     }
 
     @Override
